fix(i18n): move Compose locale provider into UI layer
