+++ app/src/main/java/com/tyranor/next/core/i18n/AppLocaleController.kt
@@ -6,12 +6,6 @@ import android.content.ContextWrapper
 import android.content.res.AssetManager
 import android.content.res.Configuration
 import android.content.res.Resources
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.remember
-import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalResources
 import com.tyranor.next.core.settings.AppSettingsStore
 import java.util.Locale
 
@@ -46,26 +40,4 @@ object AppLocaleController {
         is ContextWrapper -> findActivity(context.baseContext)
         else -> null
     }
-}
-
-@Composable
-fun ProvideAppLocale(content: @Composable () -> Unit) {
-    val baseContext = LocalContext.current
-    AppSettingsStore.initLanguage(baseContext)
-    val language = AppSettingsStore.languageState.value
-    val localizedContext = remember(baseContext, language) {
-        AppLocaleController.wrap(baseContext, language)
-    }
-    val localizedConfiguration = remember(localizedContext, language) {
-        Configuration(localizedContext.resources.configuration)
-    }
-    CompositionLocalProvider(
-        LocalContext provides localizedContext,
-        LocalConfiguration provides localizedConfiguration,
-        // stringResource() reads LocalResources rather than LocalContext. Providing
-        // the localized Resources here keeps every Compose screen in sync with the
-        // app-selected language, including sheets and dialogs opened later.
-        LocalResources provides localizedContext.resources,
-        content = content,
-    )
 }
