+++ app/src/main/java/com/tyranor/next/ui/patch/KrkrOnlinePatchActivity.kt
@@ -59,7 +59,7 @@ import com.tyranor.next.core.game.model.ScanGame
 import com.tyranor.next.core.game.model.ScanGameIntents
 import com.tyranor.next.core.settings.AppSettingsStore
 import com.tyranor.next.theme.NavWhite
-import com.tyranor.next.core.i18n.ProvideAppLocale
+import com.tyranor.next.ui.common.ProvideAppLocale
 import com.tyranor.next.theme.TyranorNextTheme
 import com.tyranor.next.ui.common.AppSearchField
 import com.tyranor.next.ui.common.TimeFormats
