+++ app/src/main/java/com/tyranor/next/ui/settings/EngineSettingsActivity.kt
@@ -13,7 +13,7 @@ import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
 import com.tyranor.next.R
 import com.tyranor.next.core.settings.AppSettingsStore
-import com.tyranor.next.core.i18n.ProvideAppLocale
+import com.tyranor.next.ui.common.ProvideAppLocale
 import com.tyranor.next.theme.TyranorNextTheme
 import com.tyranor.next.ui.common.WithoutPressIndication
