+++ app/src/main/java/com/tyranor/next/ui/settings/EngineSettingsMenuActivity.kt
@@ -39,7 +39,7 @@ import androidx.compose.ui.unit.dp
 import com.tyranor.next.R
 import com.tyranor.next.core.settings.AppSettingsStore
 import com.tyranor.next.theme.MiuixSettingsTheme
-import com.tyranor.next.core.i18n.ProvideAppLocale
+import com.tyranor.next.ui.common.ProvideAppLocale
 import com.tyranor.next.theme.TyranorNextTheme
 import com.tyranor.next.ui.common.WithoutPressIndication
 import com.tyranor.next.ui.game.startActivityWithPageTransition
